Fixes check for empty geometry in document
On diff page between two articles I got error because of wrong check for geometry in the document. Error was:
"builtins.KeyError
KeyError: 'geometry'"

F.e. on this URL:
http://localhost:6558/articles/history/716071/en
http://localhost:6558/articles/diff/716071/en/1154024/1158957

I changed it to check for geometry correctly (instead of look if geometry non-existing geometry is true I search for existing key in a document).