top navbar offset for anchors
Ensures that top fixed navbar offset is taken into account for anchors.
The solution used is to add an invisible block on top of items with an id, inspired by https://github.com/twbs/bootstrap/issues/1768#issuecomment-46519033
To limit the number of such blocks, we restrict it to h3 to h6 titles in the view-details-section part.